+++ graphql-dgs-spring-graphql/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.netflix.graphql.dgs.autoconfig.DgsSpringGraphQLAutoConfiguration
+com.netflix.graphql.dgs.springgraphql.autoconfig.DgsSpringGraphQLAutoConfiguration
